requirements.txt
Django==2.1.3
Jinja2==2.10
MarkupSafe==1.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
django-filter==2.0.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-stubs==0.2.0
drf-yasg==1.14.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
openapi-codec==1.3.2
pip==19.0.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2018.7
requests==2.20.1
ruamel.yaml==0.15.89
setuptools==40.8.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wheel==0.33.1
whitenoise==4.1.2
django-cors-headers==2.5.2